feat(p2p): shiny new peer-2-peer communication library
This one has been a loooooooong time coming. Up until this point we've been heavily focused on
getting our dataset definitions to work properly, assuming that if peers are communicating about
inaccurate datasets, well, that's not a good starting point.

In the last little while datasets have turned a corner, and we can finally turn our attention to p2p
While I don't imagine this will be anywhere near a finished state for p2p, this might be one of the
biggest steps forward this package takes.

We've reworked the request/response model into an architecture that allows methods to form fit
to the task, polling peers & sending messages until the data in question is found, or a deadline
passes, whichever comes first.

What's more, WE ACTUALLY HAVE TESTS FOR THIS STUFF. While the tests, again, have a ways to go,
coverage of this package was around 18% before now.

Before we merge this I'm hoping to ship a modification to fsrepo that caches a bunch of the info
that's flying around the network to facilitate peers covering for each other in network churn.

fuego.